Main Game Finished
- Scoring and leaderboard system complete
- Ranking icons donde
- GUID for custom songs
- Other stages backgrounds implemented
- Stage selector in editor
- Song preview start from menu and config in editor
- UI Tweaks, button highlights, etc...
- Robo floats on a hoverboard!

Things to consider adding
- Small tutorial
- Whoos sfx
- Hurt sfx

GGs ^^

+++ Assets/Scripts/Gameplay/NoteManager.cs
@@ -201,7 +201,7 @@ public class NoteManager : MonoBehaviour
 
 
         if (perfects + greats + misses != maxPossibleGrade) Debug.LogWarning($"Incoherence found between hits and max possible hits: {perfects + greats + misses}, {maxPossibleGrade}");
-        return (perfects + greats * (2 / 3)) / maxPossibleGrade * 100;
+        return ((float)perfects + (float)greats * (2f / 3f)) / maxPossibleGrade * 100f;
 
     }
